added animations

+++ components/Post.js
@@ -3,9 +3,10 @@ import { HeartIcon } from '@heroicons/react/solid'
 import React, { useEffect, useState } from 'react'
 import Moment from 'react-moment';
 import { collection, onSnapshot, doc, setDoc, deleteDoc } from 'firebase/firestore';
-import { db } from '../firebase';
-import { useSession } from 'next-auth/react';
+import { db,storage } from '../firebase';
+import { signIn, useSession } from 'next-auth/react';
 import {HeartIcon as HeartIconFilled} from "@heroicons/react/solid"
+import { deleteObject, ref } from 'firebase/storage';
 
 export default function Post({ post }) {
   const {data:session} = useSession();
@@ -23,16 +24,29 @@ useEffect(()=>{
 },[likes])
 
   async function likePost(){
-    if(hasLiked){
-      await deleteDoc(doc(db,"posts",post.id,"likes",session?.user.uid))
+    if(session){
+      if(hasLiked){
+        await deleteDoc(doc(db,"posts",post.id,"likes",session?.user.uid))
+      }
+      else{
+        await setDoc(doc(db,"posts",post.id,"likes",session?.user.uid),{
+          username: session.user.username
+        })
+      }
     }
     else{
-      await setDoc(doc(db,"posts",post.id,"likes",session?.user.uid),{
-        username: session.user.username
-      })
+      signIn()
+    }
+    
+  }
+  async function deletePost(){
+    if(window.confirm("Are you sure")){
+      deleteDoc(doc(db,"posts",post.id))
+      if(post.data().image){
+        deleteObject(ref(storage, `posts/${post.id}/image`));
+      }
     }
   }
-
 
 
   return (
@@ -63,7 +77,10 @@ useEffect(()=>{
         {/* icons */}
         <div className='flex justify-between text-gray-500 p-2'>
           <ChatIcon className='h-9 w-9 hoverEffect p-2 hover:text-sky-500 hover:bg-sky-100' />
-          <TrashIcon className='h-9 w-9 hoverEffect p-2 hover:text-red-500 hover:bg-red-100' />
+          {session?.user.uid === post?.data().id && (
+          <TrashIcon onClick={deletePost} className='h-9 w-9 hoverEffect p-2 hover:text-red-500 hover:bg-red-100' />
+          )}          
+
           <div className='flex'>
             {hasLiked?(
               <HeartIconFilled onClick={likePost}className='h-9 w-9 hoverEffect p-2 text-red-500 hover:bg-red-100' />
